feat(dw): move quota level issue states out of tabs
